Move more simplex functions to SimplexNoise unit
