+++ src/openrct2/world/map_generator/SimplexNoise.cpp
@@ -10,7 +10,11 @@
 #include "SimplexNoise.h"
 
 #include "../../util/Util.h"
+#include "HeightMap.hpp"
 #include "MapGen.h"
+#include "MapHelpers.h"
+
+#include <algorithm>
 
 namespace OpenRCT2::World::MapGenerator
 {
@@ -145,5 +149,79 @@ namespace OpenRCT2::World::MapGenerator
         float u = h < 4 ? x : y; // into 8 simple gradient directions,
         float v = h < 4 ? y : x; // and compute the dot product with (x,y).
         return ((h & 1) != 0 ? -u : u) + ((h & 2) != 0 ? -2.0f * v : 2.0f * v);
+    }
+
+    /**
+     * Smooths the height map.
+     */
+    static void smoothHeightMap(int32_t iterations, HeightMap& heightMap)
+    {
+        for (auto i = 0; i < iterations; i++)
+        {
+            auto copyHeight = heightMap;
+            for (auto y = 1; y < heightMap.height - 1; y++)
+            {
+                for (auto x = 1; x < heightMap.width - 1; x++)
+                {
+                    auto avg = 0;
+                    for (auto yy = -1; yy <= 1; yy++)
+                    {
+                        for (auto xx = -1; xx <= 1; xx++)
+                        {
+                            avg += copyHeight[{ y + yy, x + xx }];
+                        }
+                    }
+                    avg /= 9;
+                    heightMap[{ x, y }] = avg;
+                }
+            }
+        }
+    }
+
+    static void generateSimplexNoise(Settings* settings, HeightMap& heightMap)
+    {
+        float freq = settings->simplex_base_freq / 100.0f * (1.0f / heightMap.width);
+        int32_t octaves = settings->simplex_octaves;
+
+        int32_t low = settings->heightmapLow / 2;
+        int32_t high = settings->heightmapHigh / 2 - low;
+
+        NoiseRand();
+        for (int32_t y = 0; y < heightMap.height; y++)
+        {
+            for (int32_t x = 0; x < heightMap.width; x++)
+            {
+                float noiseValue = std::clamp(FractalNoise(x, y, freq, octaves, 2.0f, 0.65f), -1.0f, 1.0f);
+                float normalisedNoiseValue = (noiseValue + 1.0f) / 2.0f;
+
+                heightMap[{ x, y }] = low + static_cast<int32_t>(normalisedNoiseValue * high);
+            }
+        }
+    }
+
+    void generateSimplexMap(Settings* settings)
+    {
+        resetSurfaces(settings);
+
+        // Create the temporary height map and initialise
+        auto& mapSize = settings->mapSize;
+        auto heightMap = HeightMap(mapSize.x * 2, mapSize.y * 2);
+
+        generateSimplexNoise(settings, heightMap);
+        smoothHeightMap(2 + (UtilRand() % 6), heightMap);
+
+        // Set the game map to the height map
+        setMapHeight(settings, heightMap);
+
+        if (settings->smoothTileEdges)
+        {
+            // Set the tile slopes so that there are no cliffs
+            while (MapSmooth(1, 1, mapSize.x - 1, mapSize.y - 1))
+            {
+            }
+        }
+
+        // Add the water
+        setWaterLevel(settings->waterLevel);
     }
 } // namespace OpenRCT2::World::MapGenerator
